+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 #### HEAD
 
-- New locale
-  ([#856](https://github.com/chriso/validator.js/pull/856))
+- Added support for IP hostnames in `isEmail()`
+  ([#845](https://github.com/chriso/validator.js/pull/845))
+- Added a `no_symbols` option to `isNumeric()`
+  ([#848](https://github.com/chriso/validator.js/pull/848))
+- New and improved locales
+  ([#856](https://github.com/chriso/validator.js/pull/856),
+   [#870](https://github.com/chriso/validator.js/pull/870),
+   [#872](https://github.com/chriso/validator.js/pull/872))
 
 #### 10.4.0